GCC: Switch from ceilf to ceil

+++ src/game/Entities/StatSystem.cpp
@@ -231,7 +231,7 @@ void Unit::UpdateMaxPower(Powers power)
     value += GetModifierValue(unitMod, TOTAL_VALUE) +  bonusPower;
     value *= GetModifierValue(unitMod, TOTAL_PCT);
 
-    SetMaxPower(power, uint32(std::ceilf(value)));
+    SetMaxPower(power, uint32(std::ceil(value)));
 }
 
 void Player::UpdateAttackPowerAndDamage(bool ranged)
@@ -1019,7 +1019,7 @@ void Pet::UpdateMaxPower(Powers power)
     value += GetModifierValue(unitMod, TOTAL_VALUE) + std::max((addValue - 20) * 15 + 20, 0.f);
     value *= GetModifierValue(unitMod, TOTAL_PCT);
 
-    SetMaxPower(power, uint32(std::ceilf(value)));
+    SetMaxPower(power, uint32(std::ceil(value)));
 }
 
 void Pet::UpdateAttackPowerAndDamage(bool ranged)
